+++ pyfhirsdc/converters/activityConverter.py
@@ -2,36 +2,30 @@ import json
 
 from fhir.resources.R4B.activitydefinition import (ActivityDefinition,
                                                ActivityDefinitionDynamicValue)
+
 from fhir.resources.R4B.expression import Expression
 from fhir.resources.R4B.extension import Extension
 from fhir.resources.R4B.fhirtypes import Canonical
 
 from pyfhirsdc.config import get_defaut_fhir, get_fhir_cfg
-from pyfhirsdc.converters.utils import (clean_group_name, clean_name,
-                                        get_resource_url)
+from pyfhirsdc.converters.utils import (adv_clean_name, clean_name,
+                                        get_resource_url,get_pyfhirsdc_lib_name)
 from pyfhirsdc.serializers.json import read_resource
 
 from .extensionsConverter import append_unique
 
 
-def init_activity(filepath, id):
-    #activity_json = read_resource(filepath, "ActivityDefinition")
-    default =get_defaut_fhir('ActivityDefinition')
-    #if activity_json is not None :
-     #   activity = ActivityDefinition.parse_raw( json.dumps(activity_json))  
-    #elif default is not None:
-        # create file from default
-    activity = ActivityDefinition.parse_raw(json.dumps(default))
-    activity.id= clean_name(id)
-    activity.url=get_resource_url('ActivityDefinition',id) 
 
-    return activity
-
-def create_activity(activity_definition ,questionnaire):
-   #FIXME we should have {{context}} in questionnationnaire to define PATDOC or on the PD
+def create_activity_collect_with(questionnaire):
+    #FIXME we should have {{context}} in questionnationnaire to define PATDOC or on the PD
+    act_id = clean_name(questionnaire['id'])
+    activity_definition = ActivityDefinition.parse_raw(json.dumps(get_defaut_fhir('ActivityDefinition-collect-with')))
+    activity_definition.url=get_resource_url('ActivityDefinition',act_id) 
     activity_definition.kind = 'Task'
-    #TODO: have std approach for lif ref
-    activity_definition.library = [Canonical(get_resource_url('Library',  clean_group_name( clean_name(questionnaire['id']) ))+"|"+get_fhir_cfg().lib_version)]
+    activity_definition.id = act_id
+    activity_definition.name = questionnaire['name']
+    activity_definition.version=get_fhir_cfg().lib_version
+    activity_definition.library = [Canonical(get_resource_url('Library', get_pyfhirsdc_lib_name(act_id),True))]
     #activity_definition.useContext = [
     #  UsageContext( 
     #    code = get_code(
@@ -47,16 +41,35 @@ def create_activity(activity_definition ,questionnaire):
     #  "collect-information",
     #  "Collect information")
     # could nbe splitted into input.code / input.value
-    activity_definition.dynamicValue = [ ActivityDefinitionDynamicValue(
-          path = "focus",
-          expression = Expression(
-              language = "text/cql-identifier",
-              expression = "BackReference"
-          )
-        )
-    ]
+#{
+#    "path" : "status",
+#    "expression" : {
+#      "language" : "text/fhirpath",
+#      "expression" : "'draft' as String"
+#    }
+#  }
+    
     new_ext = Extension(
         url = "http://hl7.org/fhir/uv/cpg/StructureDefinition/cpg-collectWith",
         valueCanonical = questionnaire['url'])
-    #append_unique(activity_definition.extension, new_ext, True)
+    if activity_definition.extension is None:
+        activity_definition.extension = []
+    append_unique(activity_definition.extension, new_ext, True)
+    return activity_definition
+
+
+def create_activity_propose_diagnosis(row,library):
+    activity_definition = ActivityDefinition.parse_raw(json.dumps(get_defaut_fhir('ActivityDefinition-propose-diagnosis')))
+    activity_definition.version=get_fhir_cfg().lib_version
+    activity_definition.dynamicValue.append(
+        ActivityDefinitionDynamicValue(
+            path= "contained",
+            expression= Expression(
+            language = "text/cql-identifier",
+            expression = f"generateCondition_{row['id']}")
+        )
+    )
+    activity_definition.id= clean_name('propose-diagnosis-' + row['id'])
+    activity_definition.url=get_resource_url('ActivityDefinition',activity_definition.id)
+    activity_definition.library = [library.url+"|"+Canonical(library.version)]
     return activity_definition